requirements_ubuntu1804.txt
absl-py==0.7.1
astor==0.8.0
attrs==19.1.0
autokeras==0.4.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.6.16
chardet==3.0.4
cloudpickle==1.2.1
cycler==0.10.0
dask==2.0.0
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
gast==0.2.2
grpcio==1.22.0
h5py==2.9.0
idna==2.8
imageio==2.5.0
ipykernel==5.1.1
ipython==7.6.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.3
Jinja2==2.10.1
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.5.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
mock==3.0.5
nbconvert==5.5.0
nbformat==4.4.0
networkx==2.3
notebook==5.7.8
numpy==1.16.4
pandocfilters==1.4.2
parso==0.5.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.8.0
ptyprocess==0.6.0
Pygments==2.4.2
pyparsing==2.4.0
pyrsistent==0.14.11
python-dateutil==2.8.0
PyWavelets==1.0.3
PyYAML==5.1.1
pyzmq==18.0.0
qtconsole==4.5.1
requests==2.21.0
scikit-image==0.14.2
scikit-learn==0.20.2
scipy==1.2.0
Send2Trash==1.5.0
six==1.12.0
tensorboard==1.12.2
tensorflow-estimator==1.13.0
tensorflow-gpu==1.12.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
toolz==0.9.0
torch==1.0.1.post2
torchvision==0.2.1
tornado==6.0.3
tqdm==4.31.0
traitlets==4.3.2
urllib3==1.24.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
widgetsnbextension==3.4.2
